rework MainBar and implement MUI Swipeable Drawer for side Nav
